Fix total cycles display issues
- Fix 65535 issue by detecting uninitialized EEPROM (0xFFFFFFFF)
- Upgrade totalCycles from uint16_t to uint32_t to avoid 65535 limit
- Add validation to treat uninitialized EEPROM as 0 instead of max value
- Should resolve both the crazy number on reset and the off-by-one issue

+++ include/WebDashboard.h
@@ -64,7 +64,7 @@ private:
     CycleData cycleBuffer[MAX_CYCLE_RECORDS];
     int cycleBufferIndex;
     unsigned long lastCycleTime;
-    uint16_t totalCycles;
+    uint32_t totalCycles;  // Use 32-bit to avoid 65535 limit
     bool cycleDataLoaded;
     
     //! ********************** HOURLY DATA ************************************
